Upgrade to TinyUSB 0.21.0.
SDK 2.3.0 pins TinyUSB 0.18.0, so updating the SDK reverted the submodule
and the 0.20.0 audio API the source expected no longer existed.

audio20_control_request_t was removed upstream. Control requests now take
the control selector, channel number and entity ID out of wValue/wIndex
with TU_U16_HIGH/TU_U16_LOW, which is endian safe.

board_millis() became tusb_time_millis_api(), which the application has to
supply when no RTOS is used.

tud_init() is deprecated in favour of tusb_init(rhport, rh_init).

add_compile_definitions() takes no target argument, so the target name was
being passed as a definition. Every translation unit warned about the
invalid macro name, which drowned out two real warnings: a narrowing
conversion into audio_i2s_config_t::dma_channel and the tud_init()
deprecation.

CI: SDK 2.3.0, pico-extras sdk-2.3.0, TinyUSB 0.21.0 and current action
versions. actions/upload-release-asset has been archived since 2021, so
release packages now go up via softprops/action-gh-release.

+++ src/i2s_audio.cpp
@@ -31,7 +31,7 @@ void i2s_audio_init() {
     // We need to make sure to claim an unused channel,
     // but then audio_i2s_setup tries to claim it again...
     // So claim & unclaim it, effectively checking it's unused.
-    uint dma_channel = dma_claim_unused_channel(true);
+    uint8_t dma_channel = (uint8_t)dma_claim_unused_channel(true);
     dma_channel_unclaim(dma_channel);
 
     producer_pool = audio_new_producer_pool(&producer_format, 3, CFG_TUD_AUDIO_FUNC_1_EP_OUT_SW_BUF_SZ / sizeof(int16_t)); // todo correct size
